+++ docs/changes.rst
@@ -2,6 +2,8 @@ Release history
 ===============
 
 - feat: new attribute :attr:`Privacy.enable_phone_status_pickup_barge_in_privacy  <wxc_sdk.person_settings.privacy.Privacy>`
+- feat: new API :attr:`api.telephony.jobs.update_routing_prefix <wxc_sdk.telephony.jobs.JobsApi.update_routing_prefix>`
+
 
 1.19.0
 ------